+++ tests/gdal_geom.R
@@ -131,3 +131,7 @@ ls <- x[[1]] %>% st_cast()
 class(ls)
 
 is.na(st_bbox(ls))
+
+mp = st_combine(st_buffer(st_sfc(lapply(1:3, function(x) st_point(c(x,x)))), 0.2 * 1:3))
+plot(st_centroid(mp), add = TRUE, col = 'red') # centroid of combined geometry
+plot(st_centroid(mp, of_largest_polygon = TRUE), add = TRUE, col = 'blue', pch = 3) # center of largest sub-polygon